# Optimize UriHelper.GetRelativeUrl